e2e: util: rename ready->wait
rename the "ready" util package to "ready" and change the function
signatures for better readability.
