+++ test/e2e/util/wait/pod.go
@@ -1,10 +1,10 @@
-package ready
+package wait
 
 import (
 	corev1 "k8s.io/api/core/v1"
 )
 
-func Pod(pod corev1.Pod) bool {
+func PodReady(pod corev1.Pod) bool {
 	if pod.Status.Phase != corev1.PodRunning {
 		return false
 	}
